# backend/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Zdravdom API
  description: Home healthcare platform API
  version: 1.0.0
  contact:
    name: Zdravdom Development Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProviderStatus:
      type: string
      enum: [PENDING, VERIFIED, ACTIVE, REJECTED, SUSPENDED]
    BookingStatus:
      type: string
      enum: [REQUESTED, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED]
...
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]

    # ─── Shared ───────────────────────────────────────────────────────────────
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request data
        field:
          type: string
          example: email
        timestamp:
          type: string
          format: date-time
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          example: Home
        street:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: Slovenia
        gpsLat:
          type: number
          format: double
        gpsLng:
          type: number
          format: double
        isDefault:
          type: boolean
    EmergencyContact:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        relation:
          type: string
    InsuranceDetails:
      type: object
      properties:
        provider:
          type: string
        policyNumber:
          type: string
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        insuranceDetails:
          $ref: '#/components/schemas/InsuranceDetails'
        allergies:
          type: array
          items:
            type: string
        chronicConditions:
          type: array
          items:
            type: string
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContact'
    UpdatePatientRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        insuranceDetails:
          $ref: '#/components/schemas/InsuranceDetails'
        allergies:
          type: array
          items:
            type: string
        chronicConditions:
          type: array
          items:
            type: string
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContact'
    AddAddressRequest:
      type: object
      required:
        - label
        - street
        - city
        - postalCode
      properties:
        label:
          type: string
        street:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: Slovenia
        gpsLat:
          type: number
          format: double
        gpsLng:
          type: number
          format: double
        isDefault:
          type: boolean
          default: false
...
    GDPRExportResponse:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        downloadUrl:
          type: string
          description: Presigned S3 URL when completed

    # ─── Service / Catalog ─────────────────────────────────────────────────────
...
    ProviderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        profession:
          type: string
        specialty:
          type: string
        rating:
          type: number
          format: float
        reviewsCount:
          type: integer
        distance:
          type: number
          description: Distance in km
        languages:
          type: array
          items:
            type: string
        experience:
          type: integer
          description: Years of experience
        bio:
          type: string
        photoUrl:
          type: string
        status:
          $ref: '#/components/schemas/ProviderStatus'

    # ─── Availability ─────────────────────────────────────────────────────────
    WeeklyScheduleItem:
      type: object
      required:
        - day
        - startTime
        - endTime
      properties:
        day:
          $ref: '#/components/schemas/DayOfWeek'
        startTime:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "17:00"
    ProviderAvailability:
      type: object
      properties:
        weeklySchedule:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyScheduleItem'
        blockedDates:
          type: array
          items:
            type: string
            format: date
    UpdateAvailabilityRequest:
      type: object
      properties:
        weeklySchedule:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyScheduleItem'
        blockedDates:
          type: array
          items:
            type: string
            format: date
...
    BookingInbox:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        patientFirstName:
          type: string
        patientLastName:
          type: string
        serviceName:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        date:
          type: string
          format: date
        timeSlot:
          type: string
        status:
          $ref: '#/components/schemas/BookingStatus'
...
security:
  - BearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /patients/me:
    get:
      summary: Get current patient profile
      tags: [Patients]
      responses:
        '200':
          description: Patient profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update current patient profile
      tags: [Patients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/me/addresses:
    get:
      summary: List saved addresses
      tags: [Patients]
      responses:
        '200':
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new address
      tags: [Patients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAddressRequest'
      responses:
        '201':
          description: Address added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/me/documents:
    post:
      summary: Upload a medical document
      tags: [Patients]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - type
                - description
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  example: MEDICAL_RECORD
                description:
                  type: string
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  url:
                    type: string
                    format: uri
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/me/gdpr/export:
    get:
      summary: Trigger GDPR data export (async ZIP)
      tags: [Patients]
      responses:
        '202':
          description: Export job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GDPRExportResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Services ─────────────────────────────────────────────────────────────────
...
  /providers/me:
    get:
      summary: Get current provider profile
      tags: [Providers]
      responses:
        '200':
          description: Provider profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderSummary'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /providers/me/availability:
    get:
      summary: Get provider availability
      tags: [Providers]
      responses:
        '200':
          description: Availability schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderAvailability'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update provider availability
      tags: [Providers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAvailabilityRequest'
      responses:
        '200':
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderAvailability'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /providers/me/documents:
    post:
      summary: Upload verification documents
      tags: [Providers]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  example: LICENSE
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  url:
                    type: string
                    format: uri
                  type:
                    type: string
                  status:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /providers/inbox:
    get:
      summary: Provider's booking inbox
      tags: [Providers]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ALL, PENDING, CONFIRMED]
          description: Filter by booking status
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated booking inbox
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingInbox'
                  page:
                    type: integer
                  size:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not a provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /providers/{id}/profile:
    get:
      summary: Get public provider profile
      tags: [Providers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Provider profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderSummary'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Provider Matching ────────────────────────────────────────────────────────
  /providers:
    get:
      summary: Match providers by service, date, time and address
      tags: [Providers]
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: time
          in: query
          required: true
          schema:
            type: string
            example: "09:00"
        - name: addressId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Matched providers sorted by distance/rating
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderSummary'
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Booking Slots ───────────────────────────────────────────────────────────
...